fix: change columns types

+++ database/migrations/2020_02_28_084751_add_altitude_to_gps_points_table.php
@@ -14,7 +14,7 @@ class AddAltitudeToGpsPointsTable extends Migration
 	public function up()
 	{
 		Schema::table('gps_points', function (Blueprint $table) {
-			$table->float('altitude')
+			$table->double('altitude')
 				->after('latitude')
 				->nullable();
 		});
