+++ database/migrations/2020_02_29_184657_change_range_of_gps_fields_to_gps_points.php
@@ -14,9 +14,9 @@ class ChangeRangeOfGpsFieldsToGpsPoints extends Migration
 	public function up()
 	{
 		Schema::table('gps_points', function (Blueprint $table) {
-			$table->float('longitude')->change();
-			$table->float('latitude')->change();
-			$table->float('accuracy')->change();
+			$table->double('longitude')->change();
+			$table->double('latitude')->change();
+			$table->double('accuracy')->change();
 		});
 	}
